implement policy for Listing routes

+++ app/Http/Controllers/ListingController.php
@@ -4,7 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreListingRequest;
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 use Throwable;
 
@@ -29,15 +33,19 @@ class ListingController extends Controller
         return view('listings.show', compact('listing'));
     }
 
-    public function edit(Listing $listing): View
+    public function edit(Listing $listing): View|RedirectResponse
     {
+        // Gate::authorize('edit-listing', $listing);
+
         return view('listings.edit', compact('listing'));
     }
 
     public function store(StoreListingRequest $request): RedirectResponse
     {
         try {
-            Listing::create($request->validated());
+            Listing::create($request->validated() + [
+                'employer_id' => Auth::user()->employer->id,
+            ]);
         } catch (Throwable) {
         } finally {
             return redirect(
@@ -60,6 +68,8 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing): RedirectResponse
     {
+        Gate::authorize('edit-listing', $listing);
+
         try {
             $listing->delete();
         } catch (Throwable) {
